Factored out base serializer mixin
