+++ common/serializers/__init__.py
@@ -1,1 +1,2 @@
+from .serializer_base import *  # NOQA
 from .serializer_declarations import *  # NOQA
